src/checker.py
# -*- coding: utf-8 -*-
import logging
import json
from utils import get_cloud_sql_instances


class ComplianceChecker:
    def __init__(self, dry_run=False):
        self.dry_run = dry_run
        self.rules = self.load_rules()

    def load_rules(self):
        try:
            with open("compliance_rules/rules.json") as f:
                rules = json.load(f)
            logging.debug("Compliance rules loaded successfully.")
            return rules
        except Exception as e:
            logging.error(f"Failed to load compliance rules: {e}")
            return []

    def run_checks(self):
        for rule in self.rules:
            logging.info(f"Running check: {rule['id']} - {rule['description']}")
            check_function = getattr(self, rule["check_function"], None)
            if callable(check_function):
                check_function()
            else:
                logging.error(
                    f"Check function {rule['check_function']} not implemented."
                )

    def check_sql_instances_not_public(self):
        instances = get_cloud_sql_instances()
        for instance in instances:
            if "0.0.0.0/0" in instance.get("authorizedNetworks", []):
                logging.warning(f"Instance '{instance['name']}' is open to the world.")
                if not self.dry_run:
                    self.remediate_instance(instance)
            else:
                logging.info(f"Instance '{instance['name']}' is secure.")

    def remediate_instance(self, instance):
        logging.info(f"Remediating instance '{instance['name']}'...")
        # Placeholder for actual remediation logic
        # For example, remove '0.0.0.0/0' from authorized networks
        if self.dry_run:
            logging.debug("Dry-run mode enabled. No changes will be made.")
        else:
            logging.debug("Applying changes to the instance.")
            # Implement the API call to update the instance configuration
